+++ src/components/Header.tsx
@@ -1,8 +1,8 @@
-import React, { useState } from 'react'
+import React from 'react'
 import styled from 'styled-components/native'
 
 import { HeaderSquareIcon, HeaderStarIcon, HeaderCircleIcon, HeaderLinesIcon, LunesIcon } from '../../assets/images'
-import { reportError } from '../services/sentry'
+import SentryTestPressable from './SentryTestPressable'
 
 const Wrapper = styled.SafeAreaView`
   background-color: ${props => props.theme.colors.primary};
@@ -40,6 +40,7 @@ const VerticalLinesIcon = styled.View`
   width: 24.5px;
   height: 28px;
 `
+
 const SmileIconStyle = styled.Pressable`
   position: absolute;
   width: 80px;
@@ -49,38 +50,28 @@ const SmileIconStyle = styled.Pressable`
   top: 51px;
 `
 
-const Header = (): JSX.Element => {
-  const [counter, setCounter] = useState<number>(0)
-  const CLICKS_TO_ENABLE_SENTRY = 20
-
-  const onPress = () => {
-    if (counter > CLICKS_TO_ENABLE_SENTRY) {
-      reportError('Error for testing Sentry')
-    }
-    setCounter(oldVal => oldVal + 1)
-  }
-
-  return (
-    <Wrapper testID='header'>
-      <HeaderStyle>
-        <SquareIconStyle>
-          <HeaderSquareIcon />
-        </SquareIconStyle>
-        <StarIconStyle>
-          <HeaderStarIcon />
-        </StarIconStyle>
-        <CircleIconStyle>
-          <HeaderCircleIcon />
-        </CircleIconStyle>
-        <VerticalLinesIcon>
-          <HeaderLinesIcon />
-        </VerticalLinesIcon>
-        <SmileIconStyle onPress={onPress}>
+const Header = (): JSX.Element => (
+  <Wrapper testID='header'>
+    <HeaderStyle>
+      <SquareIconStyle>
+        <HeaderSquareIcon />
+      </SquareIconStyle>
+      <StarIconStyle>
+        <HeaderStarIcon />
+      </StarIconStyle>
+      <CircleIconStyle>
+        <HeaderCircleIcon />
+      </CircleIconStyle>
+      <VerticalLinesIcon>
+        <HeaderLinesIcon />
+      </VerticalLinesIcon>
+      <SmileIconStyle>
+        <SentryTestPressable>
           <LunesIcon />
-        </SmileIconStyle>
-      </HeaderStyle>
-    </Wrapper>
-  )
-}
+        </SentryTestPressable>
+      </SmileIconStyle>
+    </HeaderStyle>
+  </Wrapper>
+)
 
 export default Header
